license_system/requirements.txt
# 许可证管理系统依赖

# 核心依赖
Flask==2.2.3
werkzeug==2.2.3  # 明确指定与Flask 2.2.3兼容的Werkzeug版本
flask-cors==3.0.10
argon2-cffi==21.3.0
redis==4.5.1
gunicorn==20.1.0

# 数据库和安全
cryptography==39.0.1
pycryptodome==3.17.0

# 邮件
blinker==1.5
